:recycle: refactor: Enum 패키지 생성
body: LoginEnum, PetCategoryEnum, ServiceEnum 파일 옮김
