+++ app/src/main/java/com/withpet/withpet_android/enums/ServiceEnum.kt
@@ -1,4 +1,4 @@
-package com.withpet.withpet_android.others
+package com.withpet.withpet_android.enums
 
 enum class ServiceEnum {
     ENTIRE, WALK, CARE, EXP
